+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Change Log
 - [trivial] Optimize the line number calculation in vertical text orientation.
 - [dev] Update the build environment to Xcode 12 (Swift 5.3, macOS 11 SDK).
 - [dev] Replace DifferenceKit package with native CollectionDifference.
+- [dev] Update Yams from 3.0.1 to 4.0.0.
 
 
 ### Fixes