+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Change Log
 - Update Swift syntax style.
 - Update SVG syntax style.
 - Remove feature to import legacy syntax style definition files of which format was used in CotEditor 1.x.
+- [trivial] Improve tooltips of toolbar icons to reflect current document state.
 - [trivial] Optimize the line number calculation in vertical text orientation.
 - [trivial] Always enable non-contiguous layout by the normal horizontal text orientation.
 - [dev] Update the build environment to Xcode 12 (Swift 5.3, macOS 11 SDK).