+++ CHANGELOG.md
@@ -38,8 +38,9 @@ Change Log
 
 ### Fixes
 
-- Fix the progress message when highlighting a document with a multiple replacement definition.
+- Fix an issue that CotEditor Scripts didn't create a new document for output if there is no existing document window.
 - Fix an issue that “Open Scripts Folder” command in the script menu did not open the scripts folder for CotEditor but just the general Application Scripts folder.
+- Fix the progress message when highlighting a document with a multiple replacement definition.
 - Fix an issue that the preferences window sometimes performed odd resizing animation when initially display the General pane.